+++ code/Cyphers.cpp
@@ -10,14 +10,14 @@
  *	These functions are mented to help students to understand
  *	the basics of cryptography.
  *	
+ ************************************************************
+ *
+ * Functions require
+ *
+ *#include <iostream>
+ *#include <cassert>
  ************************************************************/
-
-// Comment next line for debugging purposes
-#define NDEBUG
-//
-#include <iostream>
-#include <cassert>
-
+ 
 /**
 *	Encryption function that uses Caesar cypher.
 *		This kind of cypher rotates the alphabet a designated
@@ -85,32 +85,3 @@ void CaesarDecrypt(int offst, std::string source, std::string &result){
 	}
 }
 
-
-///
-/// Test application
-///--
-///
-int main(int argc, char** argv) {
-	std::string plain, cypher, decrypted;
-	int key;
-	
-	std::cout << "\n----------------- encrypting --------------\n";
-	std::cout << "Enter plain text: ";
-	std::getline( std::cin, plain );
-	
-	std::cout << "Enter key (displacement): ";
-	std::cin >> key;
-	
-	CaesarEncrypt( key, plain, cypher );
-	
-	std::cout << "Chypher text: " << cypher;
-	
-	std::cout << "\n----------------- decrypting --------------\n";
-	std::cout << "Enter key: ";
-	std::cin >> key;
-
-	CaesarDecrypt( key, cypher, decrypted );
-	std::cout << "Decrypted: " << decrypted;
-	
-	return 0;
-}